% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_RLC_analysis.R
\name{perform_RLC_analysis}
\alias{perform_RLC_analysis}
\title{Perform RLC Analysis}
\usage{
perform_RLC_analysis(data1_l, df_x)
}
\arguments{
\item{data1_l}{Takes a data.frame with columns id, PAR, and rETR. id is the unique RLC ID. PAR is the saturating light intensity. rETR is the relative electron transport rate}

\item{df_x}{A data frame used to predict y-values using the range of light intensities in column PAR e.g data.frame(PAR = seq(0.1, 926, length = 100))}
}
\value{
A list containing p0, a plot of all the fits, and final_df, containing the extracted parameters such as Ek.
}
\description{
Perform RLC Analysis
}
